# perf(server-render): avoid unnecessary checks in `createBuffer`